backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# RabbitMQ
pika==1.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LangChain & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-community==0.0.16
langgraph==0.0.20
openai==1.10.0
anthropic==0.18.1

# Document Generation
fpdf2>=2.7.0

# Vector Store
qdrant-client==1.7.3
sentence-transformers==2.3.1

# HTTP Client
httpx==0.26.0
aiohttp==3.10.11

# RabbitMQ
aio-pika==9.4.1

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
respx==0.20.2

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
